Refactor : Infinite Carousel Correcting
