+++ Stadiumly/Stadium/custom cell/PitcherInfo.swift
